use explicit argument keywords in config and move config class to types
