+++ elixir_rems_proxy/config/__init__.py
@@ -8,23 +8,13 @@ import logging
 from pathlib import Path
 from configparser import ConfigParser
 from distutils.util import strtobool
-from typing import NamedTuple, Union
+from typing import Union
+
+from ..utils.types import Config
 
 formatting = "[%(asctime)s][%(name)s][%(process)d %(processName)s][%(levelname)-8s] (L:%(lineno)s) %(module)s | %(funcName)s: %(message)s"
 logging.basicConfig(level=logging.DEBUG if bool(strtobool(os.environ.get("DEBUG", "False"))) else logging.INFO, format=formatting)
 LOG = logging.getLogger("ELIXIR")
-
-
-class Config(NamedTuple):
-    """The app configuration."""
-
-    rems_url: str
-    repository: str
-    jwt_exp: int
-    public_key: dict
-    private_key: dict
-    host: str = "0.0.0.0"
-    port: Union[int, str] = 8080
 
 
 def load_json_file(path: Union[str, Path]) -> dict:
@@ -43,13 +33,13 @@ def parse_config_file(path: Union[str, Path]) -> Config:
     config.read(path)
 
     return Config(
-        os.environ.get("REMS_URL", config.get("rems", "rems_url")),
-        os.environ.get("GA4GH_REPOSITORY", config.get("ga4gh", "repository")),
-        int(os.environ.get("JWT_EXP", config.get("ga4gh", "jwt_exp"))),
-        load_json_file(os.environ.get("JWK_PUBLIC_KEY_FILE", Path(__file__).resolve().parent.joinpath("public_key.json"))),
-        load_json_file(os.environ.get("JWK_PRIVATE_KEY_FILE", Path(__file__).resolve().parent.joinpath("private_key.json"))),
-        os.environ.get("APP_HOST", config.get("server", "host")),
-        os.environ.get("APP_PORT", config.get("server", "port")),
+        rems_url=os.environ.get("REMS_URL", config.get("rems", "rems_url")),
+        repository=os.environ.get("GA4GH_REPOSITORY", config.get("ga4gh", "repository")),
+        jwt_exp=int(os.environ.get("JWT_EXP", config.get("ga4gh", "jwt_exp"))),
+        public_key=load_json_file(os.environ.get("JWK_PUBLIC_KEY_FILE", Path(__file__).resolve().parent.joinpath("public_key.json"))),
+        private_key=load_json_file(os.environ.get("JWK_PRIVATE_KEY_FILE", Path(__file__).resolve().parent.joinpath("private_key.json"))),
+        host=os.environ.get("APP_HOST", config.get("server", "host")),
+        port=os.environ.get("APP_PORT", config.get("server", "port")),
     )
